requirements.txt
absl-py==1.3.0
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.1
packaging==22.0
Pillow==9.3.0
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.9.3
six==1.16.0
